requirements.txt
aiofiles==24.1.0
aiosmtplib==2.0.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==3.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.2
fastapi-mail==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.8
iso8601==1.1.0
Jinja2==3.1.4
MarkupSafe==2.1.5
midtransclient==1.4.2
passlib==1.7.4
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyJWT==2.9.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
tortoise-orm==0.21.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
